Fix: 댓글 작성시 재렌더링 안되는 부분 수정
