+++ src/pages/postDetail/PostDetail.jsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import CommentBox from '../../components/comment/CommentBox';
+// import CommentBox from '../../components/comment/CommentBox';
+import Comments from '../../components/comment/Comments';
 import axios from 'axios';
 import HomePost from '../../components/PostForm/HomePost';
 import { useParams } from 'react-router-dom';
@@ -46,9 +47,7 @@ const PostDetail = () => {
                         <HomePost datas={posts.data.post} />
                     </HomePostArea>
 
-                    <CommentBox
-                        postUserName={posts.data.post.author.username}
-                    />
+                    <Comments postUserName={posts.data.post.author.username} />
                 </>
             )}
         </PageDetailArea>
